pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Анализатор страниц"
authors = ["Сергей Миргалеев <saurit.galiley@yandex.ru>"]
readme = "README.md"
repository = "https://github.com/Sgaliley/python-project-83/"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
flake8 = "^7.1.1"
psycopg2-binary = "^2.9.9"
validators = "^0.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
